backend/requirement.txt
Django==1.9.8
django-filter==1.1.0
djangorestframework==3.7.3
djangorestframework-jwt==1.11.0
MySQL-python==1.2.5
pytz==2017.3
requests==2.18.4
simplejson==3.12.0
pip install django-filter==0.9.2